Add most of our customizations to most recent version
The last time we forked formio-viewer was over a year ago.  Grabbed a new fork of it to test with.  It has changed dramatically since a year ago, so I made all of the same customizations that I could, which was frankly everything EXCEPT the on 'submissionLoad' and on 'formRender' events ... that I can't see how to do given the vast changes!